Refactored forms to be hidden behind a Form class
Things seemed to be trending in the direction of more complexity, so
I spent the last hour of the first 24 hours of Django Dash away from
the keyboard, having a quick dinner then sitting in the evening sunlight
on the back porch with coffee, a pencil, and paper.  One result was the
realization that I needed to use the Bridge Pattern here: the particular
way that forms modules are implemented needed to be hidden behind a
class, which here has indeed made other parts of the code simpler, while
reducing some redundant information that had crept in (like each form
module having to define its region, when the region can be introspected
perfectly well from the module name).