Added two PDF tax forms to the repository
This is kind of a big step, but having these forms inline will make
things as stable as possible here as we head towards the conclusion
of the Django Dash.  Plus, one form is actually smaller than all of
that CSS that I added earlier today for that dratted calendar widget!